Moved damage handling to entitybehavior
Also implemented combat overhaul compat
